chore: add userMenu outside click
