+++ src/components/admin/AdminHeader/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react';
+import React, { memo, useState } from 'react';
 import './Admin.css';
 import {
   AdminNavCategoryWrapper,
@@ -6,11 +6,7 @@ import {
   SidebarContainer,
   StyledUserName,
 } from './styled';
-import AdminSignInModal from '../../../components/common/Modal/AdminSignIn';
-import AdminSignUpModal from '../../../components/common/Modal/AdminSignUp';
-import AdminSetUserProfile from '../../../components/common/Modal/AdminSetUserProfile';
 import AdminTopMenu from '../../../components/admin/AdminTopMenu';
-import { TopMargin } from '../../../styles/layouts';
 import { useLocation } from 'react-router';
 import {
   NavInner,
@@ -25,13 +21,11 @@ import {
 } from '../../common/navigation/DeskNavigation/styled';
 import GDSCLogo from '../../../assets/GDSCLogo.svg';
 import AdminUserMenu from '../AdminUserMenu';
-import { useNavigate } from 'react-router-dom';
 import { useRecoilState } from 'recoil';
 import { localUserState } from '../../../store/localUser';
 
-const AdminHome = () => {
+const AdminHeader = () => {
   const location = useLocation();
-  const navigate = useNavigate();
   const [adminMenuHandler, setAdminMenuHandler] = useState(false);
   const [adminUser] = useRecoilState(localUserState);
   const [selectedCategory, setSelectedCategory] = useState<string>(
@@ -46,51 +40,43 @@ const AdminHome = () => {
   ];
 
   return (
-    <>
-      <AdminSignInModal />
-      <AdminSignUpModal />
-      <AdminSetUserProfile />
-      <AdminSetUserProfile />
-      <AdminNavigationWrapper>
-        <NavWrapper>
-          <NavInner>
-            <NavTaskWrapper>
-              <NavTask>
-                <StyledLogoWrapper to={'/admin'}>
-                  <StyledImg src={GDSCLogo} alt="GDSC-Chapter-Logo" />
-                  <StyledLogo>GDSC</StyledLogo>
-                  <SchoolName>Daejin </SchoolName>
-                  <SchoolNameUni>Admin</SchoolNameUni>
-                </StyledLogoWrapper>
-              </NavTask>
-            </NavTaskWrapper>
-            <AdminNavCategoryWrapper>
-              <SidebarContainer>
-                <AdminTopMenu
-                  tabs={tabs}
-                  selectedCategory={selectedCategory}
-                  setSelectedCategory={setSelectedCategory}
-                />
-              </SidebarContainer>
-              <div>
-                <AdminUserMenu
-                  isOpen={adminMenuHandler}
-                  setIsOpen={setAdminMenuHandler}
-                />
-              </div>
-              {adminUser.nickname.length > 0 && (
-                <StyledUserName
-                  onClick={() => setAdminMenuHandler(!adminMenuHandler)}
-                >
-                  Hi {adminUser.nickname}
-                </StyledUserName>
-              )}
-            </AdminNavCategoryWrapper>
-          </NavInner>
-        </NavWrapper>
-      </AdminNavigationWrapper>
-    </>
+    <AdminNavigationWrapper>
+      <NavWrapper>
+        <NavInner>
+          <NavTaskWrapper>
+            <NavTask>
+              <StyledLogoWrapper to={'/admin'}>
+                <StyledImg src={GDSCLogo} alt="GDSC-Chapter-Logo" />
+                <StyledLogo>GDSC</StyledLogo>
+                <SchoolName>Daejin </SchoolName>
+                <SchoolNameUni>Admin</SchoolNameUni>
+              </StyledLogoWrapper>
+            </NavTask>
+          </NavTaskWrapper>
+          <AdminNavCategoryWrapper>
+            <SidebarContainer>
+              <AdminTopMenu
+                tabs={tabs}
+                selectedCategory={selectedCategory}
+                setSelectedCategory={setSelectedCategory}
+              />
+            </SidebarContainer>
+            <AdminUserMenu
+              isOpen={adminMenuHandler}
+              setIsOpen={setAdminMenuHandler}
+            />
+            {adminUser.nickname.length > 0 && (
+              <StyledUserName
+                onClick={() => setAdminMenuHandler(!adminMenuHandler)}
+              >
+                Hi {adminUser.nickname}
+              </StyledUserName>
+            )}
+          </AdminNavCategoryWrapper>
+        </NavInner>
+      </NavWrapper>
+    </AdminNavigationWrapper>
   );
 };
 
-export default AdminHome;
+export default memo(AdminHeader);
