+++ src/components/admin/AdminUserMenu/styled.ts
@@ -3,16 +3,18 @@ import styled from 'styled-components';
 
 export const AdminUserMenuWrapper = styled(motion.div)`
   position: fixed;
+  right: 10px;
   top: 60px;
   justify-content: center;
+  display: flex;
+  flex-direction: column;
   text-align: center;
   background: white;
   padding: 5px 5px;
-  border: 1px solid;
-  border-color: ${(props) => props.theme.colors.grey200};
+  border: 1px solid ${({ theme }) => theme.colors.grey200};
   border-radius: 10px;
 `;
-export const MenuElement = styled(motion.div)`
+export const MenuElement = styled(motion.button)`
   font-size: 1.6rem;
   padding: 5px 15px;
   border: 0 solid;
